backend/app/requirements.txt
fastapi
uvicorn[standard]
crewai[tools]
psycopg2-binary
python-dotenv
pydantic
pyyaml